+++ init.rb
@@ -4,7 +4,7 @@ Redmine::Plugin.register :redmine_kanban do
   name        'Redmine Kanban'
   author      'ibaou'
   description 'Kanban board view for Redmine issues with drag-and-drop status transitions'
-  version     '0.1.0'
+  version     '0.2.0'
 
   requires_redmine version_or_higher: '5.0'
 
@@ -13,7 +13,7 @@ Redmine::Plugin.register :redmine_kanban do
                { kanban: [:show] },
                read: true
     permission :manage_kanban,
-               { kanban: [:update_status] },
+               { kanban: [:update_status], kanban_settings: [:update] },
                require: :member
   end
 
@@ -26,3 +26,7 @@ Redmine::Plugin.register :redmine_kanban do
 end
 
 require_relative 'lib/redmine_kanban/hooks/view_layouts_base_html_head_hook'
+require_relative 'lib/redmine_kanban/patches/projects_helper_patch'
+
+ProjectsHelper.include RedmineKanban::Patches::ProjectsHelperPatch \
+  unless ProjectsHelper.include?(RedmineKanban::Patches::ProjectsHelperPatch)